+++ frescobaldi_app/qpageview/poppler.py
@@ -62,7 +62,7 @@ class Link(link.Link):
     def __init__(self, linkobj):
         self.linkobj = linkobj
         self.area = link.Area(*linkobj.linkArea().normalized().getCoords())
-    
+
     @property
     def url(self):
         if isinstance(self.linkobj, popplerqt5.Poppler.LinkBrowse):
@@ -105,8 +105,8 @@ class PopplerPage(page.AbstractRenderedPage):
     def load(cls, filename, renderer=None):
         """Load a Poppler document, and yield of instances of this class.
 
-        The filename can also be a QByteArray or a popplerqt5.Poppler.Document 
-        instance. The specified Renderer is used, or else the global poppler 
+        The filename can also be a QByteArray or a popplerqt5.Poppler.Document
+        instance. The specified Renderer is used, or else the global poppler
         renderer.
 
         """
@@ -116,11 +116,11 @@ class PopplerPage(page.AbstractRenderedPage):
     def mutex(self):
         """No two pages of same Poppler document are rendered at the same time."""
         return self.document
-    
+
     def group(self):
         """Reimplemented to return the Poppler document our page displays a page from."""
         return self.document
-    
+
     def ident(self):
         """Reimplemented to return the page number of this page."""
         return self.pageNumber
@@ -131,7 +131,7 @@ class PopplerPage(page.AbstractRenderedPage):
         with locking.lock(self.document):
             page = self.document.page(self.pageNumber)
             return page.text(rect)
-    
+
     def links(self):
         """Reimplemented to use a different caching mechanism."""
         document, pageNumber = self.document, self.pageNumber
@@ -184,7 +184,7 @@ class PopplerRenderer(render.AbstractRenderer):
         renderBackend = printRenderBackend = 0
 
     oversampleThreshold = 96
-    
+
     def render(self, page, key, tile, paperColor=None):
         """Generate an image for the Page referred to by key."""
         if paperColor is None:
@@ -205,8 +205,8 @@ class PopplerRenderer(render.AbstractRenderer):
             key.rotation, paperColor)
         if multiplier == 2:
             image = image.scaledToWidth(tile.w, Qt.SmoothTransformation)
-        image.setDotsPerMeterX(xres * 39.37)
-        image.setDotsPerMeterY(yres * 39.37)
+        image.setDotsPerMeterX(int(xres * 39.37))
+        image.setDotsPerMeterY(int(yres * 39.37))
         return image
 
     def setRenderHints(self, doc):
